add maps for language server jump-to-x functions
Add maps/keybinds for Coc.nvim jump-to-definition, jump-to-type and
jump-to-reference functions to emulate vscode 'click-through' or jump to
definition. The 'jump to definition' is a feature of language servers and
thanks to LSP, as long as you have a language server running you can use
those functions (just have to look at what your specific language server
calls them).